+++ src/components/layout/NavComponent.vue
