+++ src/components/relays/nav/RelaysNav.vue
@@ -4,39 +4,44 @@
       <div class="flex lg:w-32 lg:px-8 lg:ml-8 sm:px-0 sm:w-0 md:w-0"></div>
       <div class="lg:flex lg:px-0">
         <div class="lg:ml-6 lg:flex lg:space-x-8">
-          <a v-for="item in store.layout.getNavGroup(this.navSlug)"
+          <router-link to="/relays/find" :class="isActive ? 'bg-color-white-100 text-white' : inactiveClass" class="inline-flex items-center mx-1 text-sm font-medium text-white">Relays</router-link>
+          <router-link to="/relays/map" class="inline-flex items-center mx-1 text-sm font-medium text-white">Map</router-link>
+          <router-link to="/relays/statistics" class="inline-flex items-center mx-1 text-sm font-medium text-white">Go to Foo</router-link>
+          <!-- <a v-for="item in store.layout.getNavGroup(this.navSlug)"
               :key="`subnav-${item.slug}`"
               :href="item.href"
               @click="setActiveContent(item.slug)"
               :class="[isActive(item) ? 'bg-gray-100 text-gray-900' : ' hover:bg-gray-50 hover:text-gray-900', 'group flex items-center px-3 text-sm font-medium']" 
               class="inline-flex items-center mx-1 text-sm font-medium text-white">
               {{ item.name }}
-          </a>
+          </a> -->
         </div>
       </div>
       <div class="width-max lg:flex lg:ml-auto">
           <RefreshComponent
-            v-bind:resultsProp="results" />
+            v-bind:resultsProp="results"
+            v-if="path == '/relays/find'"/>
         </div>
       </div>
   </div>
+
+  {{ isActive }}
 </template>
 
 <script>
 import { defineComponent, toRefs } from 'vue'
-import { items } from './config/subnav.yaml'
+import {useRoute} from 'vue-router'
+// import { items } from './config/nav.yaml'
 import { setupStore } from '@/store'
 
 import RelaysLib from '@/shared/relays-lib.js'
 import { setupNavData, mountNav, setActiveContent, loadNavContent, routeValid, parseHash, contentIsActive } from '@/shared/hash-router.js'
-import RefreshComponent from '@/components/relays/RefreshComponent.vue'
+import RefreshComponent from '@/components/relays/blocks/RefreshComponent.vue'
 
 export default defineComponent({
   title: "nostr.watch registry & network status",
   name: 'NavComponent',
   components: {
-    // PreferencesComponent,
-    // AuthComponent
     RefreshComponent,
   },
   props: {
@@ -48,7 +53,7 @@ export default defineComponent({
     },
   },
   data(){
-    return setupNavData('relays')
+    // return setupNavData('relays')
   },
   setup(props){
     const {resultsProp: results} = toRefs(props)
@@ -62,17 +67,20 @@ export default defineComponent({
   },
 
   beforeMount(){
-    this.mountNav('section', items)
+    console.log('rightt now', this.store.layout)
+    // this.mountNav('section', this.store.layout.getNavGroup('relays'))
   },
   
   mounted(){
     
   },
-  methods: Object.assign(RelaysLib, { mountNav, setActiveContent, loadNavContent, routeValid, parseHash, contentIsActive }),
+  methods: Object.assign(RelaysLib, setupNavData, mountNav, setActiveContent, loadNavContent, routeValid, parseHash, contentIsActive),
   computed: {
-    isActive(){
-        return (item) => item.slug==this.navActiveContent
-    }
+    path: function() { return useRoute().path },
+
+    // isActive(){
+    //     return (item) => item.slug==this.navActiveContent
+    // }
   },
 });
 </script>
