+++ src/components/relays/nav/config/relays.find.yaml
@@ -1,24 +1,24 @@
 items:
   - name: All
     slug: all
-    href: '#/relays/find/all'
+    href: '#all'
     description: All relays ordered by latency
   - name: Public
     slug: public
-    href: '#/relays/find/public'
+    href: '#public'
     description: All public relays ordered by latency. Public relays are relays where anyone can read or write. 
   - name: Restricted
     slug: restricted
-    href: '#/relays/find/restricted'
+    href: '#restricted'
     description: Restricted relays ordered by latency. Restricted relays are relays where there are read and/or write restrictions
   - name: Offline
     slug: offline
-    href: '#/relays/find/offline'
+    href: '#offline'
     group: relays
     description: Relays where a connection cannot currently be established
   - name: Favorites
     slug: favorite
-    href: '#/relays/find/favorite'
+    href: '#favorite'
     condition: auth
     description: Relays you have favorited
   # - name: Invalid
