+++ src/components/relays/redirects/RedirectComponent.vue
