+++ src/components/user/AuthComponent.vue
